Organalize scaffolding code; Handle errors; Refactor
